[CHORE] 코드리뷰 반영 (#7)
